requirements.txt
# 核心依赖
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# PDF处理
PyMuPDF==1.23.7
Pillow==10.1.0

# LLM服务商（按需安装）
anthropic==0.7.8
openai==1.3.7

# Web框架
flask>=3.0.0

# 工具
tqdm==4.66.1
